pyproject.toml
[project]
name = "gpu-behavioral-profiling"
version = "0.1.0"
description = "Workload-stratified GPU telemetry collection and analysis"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31",
    "pyyaml>=6.0",
    "pandas>=2.1",
    "pyarrow>=14.0",
    "scipy>=1.11",
    "scikit-learn>=1.3",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "umap-learn>=0.5",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "numpy>=1.26",
]
